requirements.txt
APScheduler==3.3.1
Babel==2.4.0
blinker==1.4
click==6.7
decorator==4.0.11
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
funcsigs==1.0.2
futures==3.1.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
MySQL-python==1.2.5
olefile==0.44
pbr==3.0.1
Pillow==4.1.1
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.10
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
tzlocal==1.4
Werkzeug==0.12.2
WTForms==2.1
XlsxWriter==0.9.6
